+++ tests/test_monergism_scrap_metadata.py
@@ -11,7 +11,9 @@ logging.basicConfig(level=logging.DEBUG)
 
 from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_metadata
 from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_general_information
-from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_get_list        
+from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_get_list  
+from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_works
+      
         
 
 def test_topic():
@@ -55,17 +57,30 @@ def test_scrap_author_general_information():
         browse_by_type = browse_by_type,
     )
     asyncio.run(ob.scrap_and_write())
+    
+    
+def test_scrap_all_author_general_information():
+    root_folder ='D:/projet_github/FOR GOD/Scraping general/test_folder'
+    browse_by_type = "speaker"
+
+    ob = monergism_scrap_general_information.MonergismScrapGeneralInformation_ALL(
+        root_folder = root_folder,
+        browse_by_type = browse_by_type,
+    )
+    #print(ob.__dict__)
+    asyncio.run(ob.download(4))
+    
 
 
 def test_get_author_all_work():
     root_folder ='D:/projet_github/FOR GOD/Scraping general/test_folder'
     browse_by_type = "speaker"
-    ob = monergism_scrap_metadata.MonergismScrapWebSiteAllAuthorTopicScripturesWork(
+    ob = monergism_scrap_works.MonergismScrapWebSiteAllAuthorTopicScripturesWork(
         root_folder,
         browse_by_type,
         overwrite_log=True
     )
-    asyncio.run(ob.download(2))
+    asyncio.run(ob.download(60))
     ob.update_downloaded_and_to_download()
     ob.write_log_file()
 
@@ -87,7 +102,7 @@ if __name__ == "__main__":
     #test_scripture()
     #test_scrap_author_general_information()
     
-    test_get_author_all_work()
+    test_scrap_all_author_general_information()
     
     
     end_time = time.time()
